requirements.txt
Django==4.0.4
django-admin==2.0.1
djangorestframework==3.13.1

